pyproject.toml
[tool.poetry]
name = "book-api"
version = "0.1.0"
description = ""
authors = ["AbdullaK123 <abdullamaseeh123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "~0.0.6"  # Changed this to be compatible with strawberry
pydantic = {extras = ["email"], version = "^2.10.4"}
alembic = "^1.14.0"
mysql-connector-python = "^9.1.0"
slowapi = "^0.1.9"
redis = "^5.2.1"
fastapi-cache2 = "^0.2.2"
asyncio = "^3.4.3"

# GraphQL dependencies
graphql-core = "3.2.3"
strawberry-graphql = {extras = ["fastapi"], version = "0.215.1"}

# Testing dependencies
pytest = "^8.3.4"
pytest-asyncio = "^0.25.1"
pytest-cov = "^6.0.0"
httpx = "^0.28.1"
boto3 = "^1.36.2"
python-dotenv = "^1.0.1"
python-magic = "^0.4.27"
moto = {extras = ["all", "ec2", "s3"], version = "^5.0.27"}
pydantic-settings = "^2.7.1"
pillow = "^11.1.0"
numpy = "^2.2.2"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.25.1"
pytest-cov = "^6.0.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
